microcloud/cmd: Reorganize CLI functions
Consolidates all the user input actions into helpers in ask.go.
Additionally adds some helpers for configuring LXD post clustering.


Squash rework
