+++ microcloud/cmd/microcloud/add.go
@@ -3,18 +3,12 @@ package main
 import (
 	"context"
 	"fmt"
-	"strings"
 
-	cephTypes "github.com/canonical/microceph/microceph/api/types"
 	"github.com/canonical/microcluster/microcluster"
-	lxdAPI "github.com/lxc/lxd/shared/api"
-	cli "github.com/lxc/lxd/shared/cmd"
-	"github.com/lxc/lxd/shared/logger"
 	"github.com/spf13/cobra"
 
 	"github.com/canonical/microcloud/microcloud/api"
 	"github.com/canonical/microcloud/microcloud/api/types"
-	"github.com/canonical/microcloud/microcloud/mdns"
 	"github.com/canonical/microcloud/microcloud/service"
 )
 
@@ -64,29 +58,9 @@ func (c *cmdAdd) Run(cmd *cobra.Command, args []string) error {
 		types.MicroOVN:  api.MicroOVNDir,
 	}
 
-	missingServices := []string{}
-	for serviceType, stateDir := range optionalServices {
-		if service.ServiceExists(serviceType, stateDir) {
-			services = append(services, serviceType)
-		} else {
-			missingServices = append(missingServices, string(serviceType))
-		}
-	}
-
-	if len(missingServices) > 0 {
-		serviceStr := strings.Join(missingServices, ",")
-		if !c.flagAuto {
-			skip, err := cli.AskBool(fmt.Sprintf("%s not found. Continue anyway? (yes/no) [default=yes]: ", serviceStr), "yes")
-			if err != nil {
-				return err
-			}
-
-			if !skip {
-				return nil
-			}
-		}
-
-		logger.Infof("Skipping %s (could not detect service state directory)", serviceStr)
+	services, err = askMissingServices(services, optionalServices, c.flagAuto)
+	if err != nil {
+		return err
 	}
 
 	s, err := service.NewServiceHandler(status.Name, addr, c.common.FlagMicroCloudDir, c.common.FlagLogDebug, c.common.FlagLogVerbose, services...)
@@ -99,81 +73,15 @@ func (c *cmdAdd) Run(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	allResources := make(map[string]*lxdAPI.Resources, len(peers)+1)
-	for peer, info := range peers {
-		allResources[peer], err = s.Services[types.LXD].(*service.LXDService).GetResources(true, peer, info.Address, info.AuthSecret)
-		if err != nil {
-			return fmt.Errorf("Failed to get system resources of peer %q: %w", peer, err)
-		}
-	}
-
-	validDisks := make(map[string][]lxdAPI.ResourcesStorageDisk, len(allResources))
-	for peer, r := range allResources {
-		validDisks[peer] = make([]lxdAPI.ResourcesStorageDisk, 0, len(r.Storage.Disks))
-		for _, disk := range r.Storage.Disks {
-			if len(disk.Partitions) == 0 {
-				validDisks[peer] = append(validDisks[peer], disk)
-			}
-		}
-	}
-
-	var diskConfig map[string][]lxdAPI.ClusterMemberConfigKey
-	var reservedDisks map[string]string
-	wantsDisks := true
-	if !c.flagAuto {
-		wantsDisks, err = cli.AskBool("Would you like to add a local LXD storage pool? (yes/no) [default=yes]: ", "yes")
-		if err != nil {
-			return err
-		}
-	}
-
-	if wantsDisks {
-		askRetry("Retry selecting disks?", c.flagAuto, func() error {
-			lxd := s.Services[types.LXD].(*service.LXDService)
-			diskConfig, reservedDisks, err = askLocalPool(peers, validDisks, c.flagAuto, c.flagWipe, *lxd)
-
-			return err
-		})
-	}
-
-	var cephDisks map[string][]cephTypes.DisksPost
-	if s.Services[types.MicroCeph] != nil {
-		ceph, ok := s.Services[types.MicroCeph].(*service.CephService)
-		if !ok {
-			return fmt.Errorf("Invalid MicroCeph service")
-		}
-
-		wantsDisks = true
-		if !c.flagAuto {
-			wantsDisks, err = cli.AskBool("Would you like to add additional disks to MicroCeph? (yes/no) [default=yes]: ", "yes")
-			if err != nil {
-				return err
-			}
-		}
-
-		if wantsDisks {
-			askRetry("Retry selecting disks?", c.flagAuto, func() error {
-				peers[status.Name] = mdns.ServerInfo{Name: status.Name, Address: addr}
-				defer delete(peers, status.Name)
-
-				cephDisks, err = askRemotePool(peers, validDisks, c.flagAuto, c.flagWipe, *ceph, reservedDisks, false)
-
-				return err
-			})
-		}
-	}
-
-	err = AddPeers(s, peers, diskConfig)
+	lxdDisks, cephDisks, err := askDisks(s, peers, false, c.flagAuto, c.flagWipe)
 	if err != nil {
 		return err
 	}
 
-	if len(cephDisks) > 0 && s.Services[types.MicroCeph] != nil {
-		_, err = finalizeRemoteDisks(*s.Services[types.MicroCeph].(*service.CephService), cephDisks, peers)
-		if err != nil {
-			return err
-		}
+	err = AddPeers(s, peers, lxdDisks)
+	if err != nil {
+		return err
 	}
 
-	return nil
+	return postClusterSetup(s, peers, lxdDisks, cephDisks)
 }
